backend/pyproject.toml
[project]
name = "podcaster-backend"
version = "0.1.0"
description = "Multi-agent live podcast platform backend"
dependencies = [
    "livekit-agents[google]>=1.3.0",
    "langgraph>=0.0.1",
    "google-genai",
    "fastapi",
    "uvicorn",
    "python-dotenv",
    "google-cloud-firestore",
    "google-cloud-storage",
    "pydantic",
]

[tool.setuptools.packages.find]
where = ["src"]
